+++ wp-content/plugins/wpforms/assets/js/admin-builder.js
@@ -1,4 +1,4 @@
-/* global wpforms_builder, wpf, jconfirm, wpforms_panel_switch, Choices, WPForms, WPFormsFormEmbedWizard, wpCookies, tinyMCE, WPFormsUtils */
+/* global wpforms_builder, wpf, jconfirm, wpforms_panel_switch, Choices, WPForms, WPFormsFormEmbedWizard, wpCookies, tinyMCE, WPFormsUtils, List */
 
 var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) {
 
@@ -72,6 +72,7 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 		 * Page load.
 		 *
 		 * @since 1.0.0
+		 * @since 1.7.9 Added `wpformsBuilderReady` hook.
 		 */
 		load: function() {
 
@@ -87,6 +88,15 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 				app.updateRevisionPreview();
 			}
 
+			// Allow callbacks to prevent making Form Builder ready...
+			const event = WPFormsUtils.triggerEvent( $builder, 'wpformsBuilderReady' );
+
+			// ...by triggering `event.preventDefault()`.
+			if ( event.isDefaultPrevented() ) {
+				return false;
+			}
+
+			// Hide loading overlay and make the Form Builder ready to use.
 			app.hideLoadingOverlay();
 
 			// Maybe display informational modal.
@@ -144,7 +154,6 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 			elements.$fieldsPreviewWrap   = $( '#wpforms-panel-fields .wpforms-panel-content-wrap' ),
 			elements.$sortableFieldsWrap  = $( '#wpforms-panel-fields .wpforms-field-wrap' );
 			elements.$addFieldsButtons    = $( '.wpforms-add-fields-button' ).not( '.not-draggable' ).not( '.warning-modal' ).not( '.education-modal' );
-
 
 			// Remove Embed button if builder opened in popup.
 			if ( app.isBuilderInPopup() ) {
@@ -243,6 +252,8 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 			};
 
 			app.dropdownField.init();
+
+			app.iconChoices.init();
 
 			app.initSomeFieldOptions();
 
@@ -450,7 +461,22 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 				 */
 				choicesInit: function( $element ) {
 
-					var instance = new Choices( $element[0], app.dropdownField.config.args );
+					let useAjax = $element.data( 'choicesjs-use-ajax' ) === 1,
+						instance;
+
+					if ( $element.data( 'choicesjs-callback-fn' ) === 'select_pages' ) {
+
+						instance = WPForms.Admin.Builder.WPFormsChoicesJS.setup(
+							$element[0],
+							app.dropdownField.config.args,
+							{
+								action: 'wpforms_ajax_search_pages_for_dropdown',
+								nonce: useAjax ? wpforms_builder.nonce : null,
+							}
+						);
+					} else {
+						instance = new Choices( $element[0], app.dropdownField.config.args );
+					}
 
 					app.dropdownField.helpers.setInstance( $element, instance );
 					app.dropdownField.helpers.addPlaceholderChoice( $element, instance );
@@ -616,8 +642,12 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 				 * @returns {boolean} False if a fake placeholder wasn't added.
 				 */
 				addPlaceholderChoice: function( $jquerySelector, instance ) {
+					const wpFormsField = $jquerySelector.closest( '.wpforms-field' );
+					if ( wpFormsField.length <= 0 ) {
+						return false;
+					}
 
-					var fieldId     = $jquerySelector.closest( '.wpforms-field' ).data().fieldId,
+					var fieldId     = wpFormsField.data().fieldId,
 						hasDefaults = app.dropdownField.helpers.hasDefaults( fieldId );
 
 					if ( app.dropdownField.helpers.isDynamicChoices( fieldId ) ) {
@@ -1731,7 +1761,13 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 					fieldID       = $optionRow.data( 'field-id' ),
 					$fieldOptions = $( '#wpforms-field-option-' + fieldID ),
 					checked       = $this.is( ':checked' ),
-					type          = $fieldOptions.find( '.wpforms-field-option-hidden-type' ).val();
+					type          = $fieldOptions.find( '.wpforms-field-option-hidden-type' ).val(),
+					$iconToggle  = $optionRow.siblings( '.wpforms-field-option-row-choices_icons' ).find( 'input' );
+
+				// Toggle icon choices off.
+				if ( checked && $iconToggle.is( ':checked' ) ) {
+					$iconToggle.prop( 'checked', false ).trigger( 'change' );
+				}
 
 				$optionRow.find( '.wpforms-alert' ).toggleClass( 'wpforms-hidden' );
 				$fieldOptions.find( '.wpforms-field-option-row-choices ul' ).toggleClass( 'show-images' );
@@ -2319,7 +2355,7 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 					value = $this.val(),
 					id = $this.parent().data( 'field-id' );
 				if ( value ) {
-					$( '#wpforms-field-' + id ).find( '.wpforms-pagebreak-title' ).text( '(' + value + ')' );
+					$( '#wpforms-field-' + id ).find( '.wpforms-pagebreak-title' ).text( value );
 				} else {
 					$( '#wpforms-field-' + id ).find( '.wpforms-pagebreak-title' ).empty();
 				}
@@ -2495,11 +2531,10 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 			$( document ).on( 'input', '.wpforms-field-option-row-icon_color input.wpforms-color-picker', function() {
 
 				var $this     = $( this ),
-					value     = $this.val(),
 					id        = $this.closest( '.wpforms-field-option-row' ).data( 'field-id' ),
 					$icons    = $( '#wpforms-field-' + id + ' > i.fa' );
 
-				$icons.css( 'color', value );
+				$icons.css( 'color', app.getValidColorPickerValue( $this ) );
 			} );
 
 			// Real-time updates for Checkbox field Disclaimer option.
@@ -2553,8 +2588,9 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 			// Real-time updates for Number Slider field.
 			app.numberSliderEvents( $builder );
 
-			// Hide image choices if dynamic choices is not off.
+			// Hide image and icon choices if dynamic choices is not off.
 			app.fieldDynamicChoiceToggleImageChoices();
+			app.fieldDynamicChoiceToggleIconChoices();
 
 			// Real-time updates for Payment field's 'Show price after item label' option.
 			$builder.on( 'change', '.wpforms-field-option-row-show_price_after_labels input', function( e ) {
@@ -3221,6 +3257,10 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 						btnClass: 'btn-confirm',
 						keys: [ 'enter' ],
 						action: function() {
+
+							// Disable the current button to avoid firing multiple click events.
+							// By default, "jconfirm" tends to destroy any modal DOM element upon button click.
+							this.$$confirm.prop( 'disabled', true );
 
 							const beforeEvent = WPFormsUtils.triggerEvent( $builder, 'wpformsBeforeFieldDuplicate', [ id, $field ]  );
 
@@ -4004,7 +4044,11 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 			$choice.attr( 'data-key', id );
 			$choice.find( 'input.label' ).val( '' ).attr( 'name', 'fields[' + fieldID + '][choices][' + id + '][label]' );
 			$choice.find( 'input.value' ).val( '' ).attr( 'name', 'fields[' + fieldID + '][choices][' + id + '][value]' );
-			$choice.find( 'input.source' ).val( '' ).attr( 'name', 'fields[' + fieldID + '][choices][' + id + '][image]' );
+			$choice.find( '.wpforms-image-upload input.source' ).val( '' ).attr( 'name', 'fields[' + fieldID + '][choices][' + id + '][image]' );
+			$choice.find( '.wpforms-icon-select input.source-icon' ).val( wpforms_builder.icon_choices.default_icon ).attr( 'name', 'fields[' + fieldID + '][choices][' + id + '][icon]' );
+			$choice.find( '.wpforms-icon-select input.source-icon-style' ).val( wpforms_builder.icon_choices.default_icon_style ).attr( 'name', 'fields[' + fieldID + '][choices][' + id + '][icon_style]' );
+			$choice.find( '.wpforms-icon-select .ic-fa-preview' ).removeClass().addClass( 'ic-fa-preview ic-fa-' + wpforms_builder.icon_choices.default_icon_style + ' ic-fa-' + wpforms_builder.icon_choices.default_icon );
+			$choice.find( '.wpforms-icon-select .ic-fa-preview + span' ).text( wpforms_builder.icon_choices.default_icon );
 			$choice.find( 'input.default' ).attr( 'name', 'fields[' + fieldID + '][choices][' + id + '][default]' ).prop( 'checked', false );
 			$choice.find( '.preview' ).empty();
 			$choice.find( '.wpforms-image-upload-add' ).show();
@@ -4174,6 +4218,12 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 						order:    slicedOrder,
 						type:     'radio',
 					};
+
+				// If Icon Choices is on, get the valid color.
+				if ( fieldSettings.choices_icons ) {
+					// eslint-disable-next-line camelcase
+					data.settings.choices_icons_color = app.getValidColorPickerValue( $( '#wpforms-field-option-' + id + '-choices_icons_color' ) );
+				}
 
 				// Slice choices for preview.
 				slicedOrder.forEach( function( entry ) {
@@ -4361,6 +4411,11 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 
 			$this.prop( 'disabled', true ).html( $this.html() + ' ' + s.spinner );
 			$choice.find( 'input.value,input.label' ).attr( 'value', '' );
+			$choice.find( 'input.default' ).attr( 'checked', false );
+			$choice.find( 'input.source-icon' ).attr( 'value', wpforms_builder.icon_choices.default_icon );
+			$choice.find( 'input.source-icon-style' ).attr( 'value', wpforms_builder.icon_choices.default_icon_style );
+			$choice.find( '.ic-fa-preview' ).removeClass().addClass( `ic-fa-preview ic-fa-${wpforms_builder.icon_choices.default_icon_style} ic-fa-${wpforms_builder.icon_choices.default_icon}` );
+			$choice.find( '.ic-fa-preview + span' ).text( wpforms_builder.icon_choices.default_icon );
 			choice = $choice.html();
 
 			for ( var key in newValues ) {
@@ -4576,10 +4631,12 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 				type        = $( '#wpforms-field-option-' + id ).find( '.wpforms-field-option-hidden-type' ).val(),
 				$field      = $( '#wpforms-field-' + id ),
 				$choices    = $( '#wpforms-field-option-row-' + id + '-choices' ),
-				$images     = $( '#wpforms-field-option-' + id + '-choices_images' );
+				$images     = $( '#wpforms-field-option-' + id + '-choices_images' ),
+				$icons      = $( '#wpforms-field-option-' + id + '-choices_icons' );
 
-			// Hide image choices if dynamic choices is not off.
+			// Hide image and icon choices if dynamic choices is not off.
 			app.fieldDynamicChoiceToggleImageChoices();
+			app.fieldDynamicChoiceToggleIconChoices();
 
 			// Fire an event when a field's dynamic choices option was changed.
 			$builder.trigger( 'wpformsFieldDynamicChoiceToggle' );
@@ -4596,8 +4653,9 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 			 */
 			if ( '' !== value ) {
 
-				// Hide choice images option, not applicable.
+				// Hide choice images and icons options, not applicable.
 				$images.addClass( 'wpforms-hidden' );
+				$icons.addClass( 'wpforms-hidden' );
 
 				// Hide `Bulk Add` toggle.
 				$choices.find( '.toggle-bulk-add-display' ).addClass( 'wpforms-hidden' );
@@ -4640,8 +4698,9 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 			 * "Off" - no dynamic populating.
 			 */
 
-			// Show choice images option.
+			// Show choice images and icons options.
 			$images.removeClass( 'wpforms-hidden' );
+			$icons.removeClass( 'wpforms-hidden' );
 
 			// Show `Bulk Add` toggle.
 			$choices.find( '.toggle-bulk-add-display' ).removeClass( 'wpforms-hidden' );
@@ -4929,6 +4988,38 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 				if ( ! isImageChoices || isDynamicChoices ) {
 					$option
 						.find( '.wpforms-field-option-row-choices_images_style' )
+						.addClass( 'wpforms-hidden' );
+				}
+			} );
+		},
+
+		/**
+		 * Hide icon choice toggle, icon choices color, size and style options if Dynamic choices is not OFF.
+		 *
+		 * @since 1.7.9
+		 */
+		fieldDynamicChoiceToggleIconChoices: function() {
+
+			$( '#wpforms-builder .wpforms-field-options .wpforms-field-option' ).each( function( key, value ) {
+
+				const $option        = $( value ),
+					dynamicChoices   = $option.find( '.wpforms-field-option-row-dynamic_choices select' ).val(),
+					isDynamicChoices = typeof dynamicChoices !== 'undefined' && '' !== dynamicChoices,
+					isIconChoices    = $option.find( '.wpforms-field-option-row-choices_icons input' ).is( ':checked' );
+
+				$option
+					.find( '.wpforms-field-option-row-choices_icons' )
+					.toggleClass( 'wpforms-hidden', isDynamicChoices );
+
+				if ( ! isIconChoices || isDynamicChoices ) {
+					$option
+						.find( '.wpforms-field-option-row-choices_icons_color' )
+						.addClass( 'wpforms-hidden' );
+					$option
+						.find( '.wpforms-field-option-row-choices_icons_size' )
+						.addClass( 'wpforms-hidden' );
+					$option
+						.find( '.wpforms-field-option-row-choices_icons_style' )
 						.addClass( 'wpforms-hidden' );
 				}
 			} );
@@ -5722,6 +5813,7 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 
 								// Reset the confirmation type to the 1st one.
 								if ( blockType === 'confirmation' ) {
+									app.prepareChoicesJSField( $addedSettingBlock, nextID );
 									app.confirmationFieldsToggle( $( '.wpforms-panel-field-confirmations-type' ).first() );
 								}
 
@@ -5754,6 +5846,48 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 					$( modal.buttons.confirm.el ).trigger( 'click' );
 				}
 			} );
+		},
+
+		/**
+		 * Reset the 'Select Page' field to it's initial state then
+		 * re-initialize ChoicesJS on it.
+		 *
+		 * @since 1.7.9
+		 *
+		 * @param {jQuery} $addedSettingBlock  Newly added Settings Block jQuery object.
+		 * @param {number} addedSettingBlockID Number ID used when `$addedSettingBlock` was created.
+		 */
+		prepareChoicesJSField: function( $addedSettingBlock, addedSettingBlockID ) {
+
+			const $addedConfirmationWrap = $addedSettingBlock.find( `#wpforms-panel-field-confirmations-${addedSettingBlockID}-page-wrap` );
+			if ( $addedConfirmationWrap.length <= 0 ) {
+				return;
+			}
+
+			const $confirmationSelectPageField = $addedConfirmationWrap.find( `#wpforms-panel-field-confirmations-${addedSettingBlockID}-page` );
+			if ( $confirmationSelectPageField.length <= 0 && ! $confirmationSelectPageField.hasClass( 'choicesjs-select' ) ) {
+				return;
+			}
+
+			const $choicesWrapper = $addedConfirmationWrap.find( '.choices' );
+			if ( $choicesWrapper.length <= 0 ) {
+				return;
+			}
+
+			// Remove ChoicesJS-related attr.
+			const $selectPageField = $confirmationSelectPageField.first();
+			$selectPageField.removeAttr( 'data-choice' );
+			$selectPageField.removeAttr( 'hidden' );
+			$selectPageField.removeClass( 'choices__input' );
+
+			// Move the select page field to it's initial location in the DOM.
+			$( $selectPageField ).appendTo( $addedConfirmationWrap.first() );
+
+			// Remove the `.choices` wrapper.
+			$choicesWrapper.first().remove();
+
+			// Re-init ChoicesJS.
+			app.dropdownField.events.choicesInit( $selectPageField );
 		},
 
 		/**
@@ -6404,7 +6538,6 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 
 			closeConfirmation = ! ! confirm;
 		},
-
 
 		/**
 		 * Check current form state.
@@ -7287,8 +7420,544 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 		},
 
 		//--------------------------------------------------------------------//
+		// Icon Choices
+		//--------------------------------------------------------------------//
+
+		/**
+		 * Icon Choices component.
+		 *
+		 * @since 1.7.9
+		 */
+		iconChoices: {
+
+			/**
+			 * Runtime component cache.
+			 *
+			 * toggle: "Use icon choices" toggle that initiated the installation.
+			 * previousModal: Last open modal that may need to be closed.
+			 *
+			 * @since 1.7.9
+			 */
+			cache: {},
+
+			/**
+			 * Component configuration settings.
+			 *
+			 * @since 1.7.9
+			 */
+			config: {
+				colorPropertyName: '--wpforms-icon-choices-color',
+			},
+
+			/**
+			 * Initialize the component.
+			 *
+			 * @since 1.7.9
+			 */
+			init: function() {
+
+				// Extend jquery-confirm plugin with max-height support for the content area.
+				app.iconChoices.extendJqueryConfirm();
+
+				$builder.on( 'wpformsBuilderReady', function( event ) {
+
+					// If there are Icon Choices fields but the library is not installed - force install prompt.
+					if ( wpforms_builder.icon_choices.is_active && ! wpforms_builder.icon_choices.is_installed ) {
+
+						app.iconChoices.openInstallPromptModal( true );
+
+						// Prevent the Form Builder from getting ready (hold the loading state).
+						event.preventDefault();
+					}
+				} );
+
+				// Toggle Icon Choices on or off.
+				$builder.on( 'change', '.wpforms-field-option-row-choices_icons input', app.iconChoices.toggleIconChoices );
+
+				// Change accent color.
+				$builder.on( 'change', '.wpforms-field-option-row-choices_icons_color .wpforms-color-picker', app.iconChoices.changeIconsColor );
+
+				// Just update field preview when option value is changed (style, size).
+				$builder.on( 'change', '.wpforms-field-option-row-choices_icons_style select, .wpforms-field-option-row-choices_icons_size select', function() {
+
+					const fieldID = $( this ).parent().data( 'field-id' ),
+						fieldType = $( '#wpforms-field-option-' + fieldID ).find( '.wpforms-field-option-hidden-type' ).val();
+
+					app.fieldChoiceUpdate( fieldType, fieldID );
+				} );
+
+				// Open Icon Picker modal.
+				$builder.on( 'click', '.wpforms-field-option-row-choices .choices-list .wpforms-icon-select', app.iconChoices.openIconPickerModal );
+			},
+
+			/**
+			 * Turn the feature on or off.
+			 *
+			 * @since 1.7.9
+			 */
+			toggleIconChoices: function() { // eslint-disable-line complexity
+
+				const $this = $( this ),
+					checked = $this.is( ':checked' ),
+					fieldID = $this.closest( '.wpforms-field-option-row' ).data( 'field-id' );
+
+				// Check if required icon library is installed.
+				if ( checked && ! wpforms_builder.icon_choices.is_installed ) {
+
+					app.iconChoices.cache.toggle = $this;
+					app.iconChoices.openInstallPromptModal();
+
+					return;
+				}
+
+				const $fieldOptions = $( `#wpforms-field-option-${fieldID}` ),
+					$imageChoices   = $fieldOptions.find( `#wpforms-field-option-${fieldID}-choices_images` ),
+					$choicesList    = $fieldOptions.find( `#wpforms-field-option-row-${fieldID}-choices ul` );
+
+				// Turn Image Choice off.
+				if ( checked && $imageChoices.is( ':checked' ) ) {
+					$imageChoices.prop( 'checked', false ).trigger( 'change' );
+				}
+
+				// Toggle Advanced > Dynamic Choices on or off.
+				$fieldOptions.find( `#wpforms-field-option-row-${fieldID}-dynamic_choices` ).toggleClass( 'wpforms-hidden', checked );
+
+				// Toggle subfields.
+				$fieldOptions.find( `#wpforms-field-option-row-${fieldID}-choices_icons_color` ).toggleClass( 'wpforms-hidden' );
+				$fieldOptions.find( `#wpforms-field-option-row-${fieldID}-choices_icons_size` ).toggleClass( 'wpforms-hidden' );
+				$fieldOptions.find( `#wpforms-field-option-row-${fieldID}-choices_icons_style` ).toggleClass( 'wpforms-hidden' );
+
+				const $colorOption = $fieldOptions.find( `#wpforms-field-option-${fieldID}-choices_icons_color` ),
+					colorValue     = _.isEmpty( $colorOption.val() ) ? wpforms_builder.icon_choices.default_color : $colorOption.val();
+
+				// Set accent color for all choices.
+				$choicesList.prop( 'style', `${app.iconChoices.config.colorPropertyName}: ${colorValue};` );
+
+				// Toggle icon selectors with previews for all choices.
+				$choicesList.toggleClass( 'show-icons', checked );
+
+				// Set layout to inline on activation, revert to one column on deactivation.
+				$fieldOptions.find( `#wpforms-field-option-${fieldID}-input_columns` ).val( checked ? 'inline' : '' ).trigger( 'change' );
+
+				// Finally, update the preview.
+				app.fieldChoiceUpdate( $fieldOptions.find( '.wpforms-field-option-hidden-type' ).val(), fieldID );
+			},
+
+			/**
+			 * Change accent color and update previews.
+			 *
+			 * @since 1.7.9
+			 */
+			changeIconsColor: function() {
+
+				const $this       = $( this ),
+					fieldID       = $this.parents( '.wpforms-field-option-row' ).data( 'field-id' ),
+					$field        = $( '#wpforms-field-option-' + fieldID ),
+					type          = $field.find( '.wpforms-field-option-hidden-type' ).val(),
+					$choicesList  = $field.find( '.wpforms-field-option-row-choices .choices-list' ),
+					colorValue    = app.getValidColorPickerValue( $this );
+
+				// Update icons color in options panel.
+				$choicesList.prop( 'style', `${app.iconChoices.config.colorPropertyName}: ${colorValue};` );
+
+				// Update preview.
+				app.fieldChoiceUpdate( type, fieldID );
+			},
+
+			/**
+			 * Open a modal prompting to install the icon library for Icon Choices.
+			 *
+			 * @since 1.7.9
+			 *
+			 * @param {boolean} force Whether it's a normal installation procedure or forced if the library is needed but is missing.
+			 */
+			openInstallPromptModal: function( force = false ) {
+
+				const content = force ?
+					wpforms_builder.icon_choices.strings.reinstall_prompt_content :
+					wpforms_builder.icon_choices.strings.install_prompt_content;
+
+				const modal = $.confirm( {
+					title: wpforms_builder.heads_up,
+					content: content,
+					icon: 'fa fa-info-circle',
+					type: 'orange',
+					buttons: {
+						continue: {
+							text: wpforms_builder.continue,
+							btnClass: 'btn-confirm',
+							keys: [ 'enter' ],
+							action: function() {
+
+								this.setIcon( 'fa fa-cloud-download' );
+								this.setTitle( wpforms_builder.icon_choices.strings.install_title );
+								this.setContent( wpforms_builder.icon_choices.strings.install_content );
+
+								$.each( this.buttons, function( _index, button ) {
+									button.hide();
+								} );
+
+								app.iconChoices.installIconLibrary();
+
+								// Do not close the modal.
+								return false;
+							},
+						},
+					},
+					onOpen: function() {
+
+						// Turn the toggle off during normal installation.
+						if ( ! force && app.iconChoices.cache.toggle ) {
+							app.iconChoices.cache.toggle.prop( 'checked', false );
+						}
+
+						app.iconChoices.cache.previousModal = this;
+					},
+				} );
+
+				// Add a Cancel button for normal installation routine only.
+				if ( ! force ) {
+					modal.buttons.cancel = {
+						text: wpforms_builder.cancel,
+						keys: [ 'esc' ],
+						action: function() {
+
+							app.iconChoices.cache.toggle.prop( 'checked', false );
+						},
+					};
+				}
+			},
+
+			/**
+			 * Silently download and install the icon library on the server.
+			 *
+			 * @since 1.7.9
+			 */
+			installIconLibrary: function() {
+
+				const data = {
+					'nonce': wpforms_builder.nonce,
+					'action': 'wpforms_icon_choices_install',
+				};
+
+				$.ajaxSetup( {
+					type: 'POST',
+					timeout: 120000, // 2 minutes.
+				} );
+
+				$.post( wpforms_builder.ajax_url, data, function( response ) {
+
+					response.success ?
+						app.iconChoices.openInstallSuccessModal() :
+						app.iconChoices.openInstallErrorModal( response );
+
+				} ).fail( function( jqXHR ) {
+
+					app.iconChoices.openInstallErrorModal( jqXHR );
+				} );
+			},
+
+			/**
+			 * Open a modal on icon library installation success.
+			 *
+			 * @since 1.7.9
+			 */
+			openInstallSuccessModal: function() {
+
+				$.confirm( {
+					title: wpforms_builder.done,
+					content: wpforms_builder.icon_choices.strings.install_success_content,
+					icon: 'fa fa-check-circle',
+					type: 'green',
+					buttons: {
+						confirm: {
+							text: wpforms_builder.ok,
+							btnClass: 'btn-confirm',
+							keys: [ 'enter' ],
+							action: function() {
+
+								if ( app.iconChoices.cache.toggle ) {
+									app.iconChoices.cache.toggle.prop( 'checked', true );
+
+									const fieldId = app.iconChoices.cache.toggle.parents( '.wpforms-field-option-row' ).data( 'field-id' );
+									const $imageChoices = $builder.find( `#wpforms-field-option-${fieldId}-choices_images` );
+
+									// Turn Image Choice off, if needed, without triggering change event.
+									if ( $imageChoices.is( ':checked' ) ) {
+										$imageChoices.prop( 'checked', false );
+									}
+								}
+
+								app.formSave( false ).done( function() {
+
+									window.location.reload();
+								} );
+							},
+						},
+					},
+					onOpen: function() {
+
+						if ( app.iconChoices.cache.toggle ) {
+							const fieldId = app.iconChoices.cache.toggle.parents( '.wpforms-field-option-row-choices_icons' ).data( 'field-id' );
+
+							$builder.find( `#wpforms-field-option-${fieldId}-input_columns` ).val( 'inline' );
+						}
+
+						app.iconChoices.cache.previousModal.close();
+					},
+				} );
+			},
+
+			/**
+			 * Open a modal on icon library installation failure.
+			 *
+			 * @since 1.7.9
+			 *
+			 * @param {object} errorData Unsuccessful ajax JSON response or jqXHR object.
+			 */
+			openInstallErrorModal: function( errorData ) {
+
+				$.confirm( {
+					title: wpforms_builder.uh_oh,
+					content: wpforms_builder.icon_choices.strings.install_error_content,
+					icon: 'fa fa-exclamation-circle',
+					type: 'red',
+					buttons: {
+						confirm: {
+							text: wpforms_builder.ok,
+							btnClass: 'btn-confirm',
+							keys: [ 'enter' ],
+							action: function() {
+
+								if ( app.iconChoices.cache.toggle ) {
+									app.iconChoices.cache.toggle.prop( 'checked', false );
+								} else {
+									app.formSaveError();
+								}
+							},
+						},
+					},
+					onOpen: function() {
+
+						wpf.debug( errorData );
+						app.iconChoices.cache.previousModal.close();
+					},
+					onDestroy: function() {
+
+						// Clean up the cache, we're done.
+						delete app.iconChoices.cache.previousModal;
+						delete app.iconChoices.cache.toggle;
+					},
+				} );
+			},
+
+			/**
+			 * Extend jquery-confirm plugin with support of max-height for the content area.
+			 *
+			 * @since 1.7.9
+			 */
+			extendJqueryConfirm: function() {
+
+				// Extend a method of global instance.
+				window.Jconfirm.prototype._updateContentMaxHeight = function() {
+
+					const height = $( window ).height() - ( this.$jconfirmBox.outerHeight() - this.$contentPane.outerHeight() ) - ( this.offsetTop + this.offsetBottom );
+
+					// Custom property, if set via jquery-confirm options.
+					const maxHeight = this.contentMaxHeight || height;
+
+					this.$contentPane.css( {
+						'max-height': Math.min( maxHeight, height ) + 'px',
+					} );
+				};
+			},
+
+			/**
+			 * Open Icon Picker modal.
+			 *
+			 * @since 1.7.9
+			 */
+			openIconPickerModal: function() {
+
+				const $this = $( this );
+
+				const data = {
+					fieldId:           $this.parents( '.wpforms-field-option-row' ).data( 'field-id' ),
+					choiceId:          $this.parent().data( 'key' ),
+					selectedIcon:      $this.find( '.source-icon' ).val(),
+					selectedIconStyle: $this.find( '.source-icon-style' ).val(),
+				};
+
+				const title = `
+					${wpforms_builder.icon_choices.strings.icon_picker_title}
+					<span class="wpforms-icon-picker-description">${wpforms_builder.icon_choices.strings.icon_picker_description}</span>
+					<input type="text" placeholder="${wpforms_builder.icon_choices.strings.icon_picker_search_placeholder}" class="search" id="wpforms-icon-picker-search">
+				`;
+
+				const content = `
+					<div class="wpforms-icon-picker-container" id="wpforms-icon-picker-icons">
+						<ul class="wpforms-icon-picker-icons" data-field-id="${data.fieldId}" data-choice-id="${data.choiceId}"></ul>
+						<ul class="wpforms-icon-picker-pagination"></ul>
+						<p class="wpforms-icon-picker-not-found wpforms-hidden" data-message="${wpforms_builder.icon_choices.strings.icon_picker_not_found}"></>
+					</div>`;
+
+				$.confirm( {
+					title: title,
+					titleClass: 'wpforms-icon-picker-title',
+					content: content,
+					icon: false,
+					closeIcon: true,
+					type: 'orange',
+					backgroundDismiss: true,
+					boxWidth: 800,
+					contentMaxHeight: 368, // Custom property, see app.iconChoices.extendJqueryConfirm().
+					smoothContent: false,
+					buttons: false,
+					onOpenBefore: function() {
+
+						// Add custom classes to target various elements.
+						this.$body.addClass( 'wpforms-icon-picker-jconfirm-box' );
+						this.$contentPane.addClass( 'wpforms-icon-picker-jconfirm-content-pane' );
+					},
+					onContentReady: function() {
+
+						const modal = this;
+
+						// Initialize the list of icons with List.js and display 1st page.
+						app.iconChoices.initIconsList( data );
+
+						// Focus the search input.
+						modal.$title.find( '.search' ).focus();
+
+						// Listen for clicks on icons to selected them.
+						modal.$content.find( '.wpforms-icon-picker-icons' ).on( 'click', 'li', function() {
+
+							app.iconChoices.selectIcon( modal, $( this ) );
+						} );
+					},
+				} );
+			},
+
+			/**
+			 * Initialize List.js in the Icon Selector modal on demand and cache it.
+			 *
+			 * @since 1.7.9
+			 *
+			 * @param {object} data Source option data - field and choice IDs, selected icon name and style.
+			 */
+			initIconsList: function( data ) {
+
+				const options = {
+					valueNames: [ 'name' ],
+					listClass: 'wpforms-icon-picker-icons',
+					page: wpforms_builder.icon_choices.icons_per_page,
+					pagination: {
+						paginationClass: 'wpforms-icon-picker-pagination',
+					},
+					item: function( values ) {
+
+						const maybeSelectedClass = ( values.icon === data.selectedIcon && values.style === data.selectedIconStyle ) ? 'class="selected"' : '';
+
+						return `
+								<li data-icon="${values.icon}" data-icon-style="${values.style}"${maybeSelectedClass}>
+									<i class="ic-fa-${values.style} ic-fa-${values.icon}"></i>
+									<span class="name">${values.icon}</span>
+								</li>`;
+					},
+					indexAsync: true,
+				};
+
+				// Initialize List.js instance.
+				const iconsList = new List( 'wpforms-icon-picker-icons', options, wpforms_builder.icon_choices.icons );
+
+				// Initialize infinite scroll pagination on the list instance.
+				app.iconChoices.infiniteScrollPagination( iconsList );
+
+				// Bind search to custom input.
+				$( '#wpforms-icon-picker-search' ).on( 'keyup', function() {
+
+					// Custom partial match search.
+					iconsList.search( $( this ).val(), [ 'name' ], function( searchString, columns ) {
+						for ( let index = 0, length = iconsList.items.length; index < length; index++ ) {
+							iconsList.items[index].found = ( new RegExp( searchString ) ).test( iconsList.items[index].values().icon );
+						}
+					} );
+				} );
+
+				// Show "nothing found" message if search returned no results.
+				iconsList.on( 'searchComplete', function() {
+
+					const $element = $( '.wpforms-icon-picker-not-found' );
+
+					$element.html( $element.data( 'message' ).replace( '{keyword}', $( '#wpforms-icon-picker-search' ).val() ) );
+					$element.toggleClass( 'wpforms-hidden', ! _.isEmpty( iconsList.matchingItems ) );
+				} );
+			},
+
+			/**
+			 * Handle infinite scroll on the list of icons.
+			 *
+			 * @since 1.7.9
+			 *
+			 * @param {object} list List.js instance.
+			 */
+			infiniteScrollPagination: function( list ) {
+
+				let page = 1;
+
+				const options = {
+					root: document.querySelector( '.wpforms-icon-picker-jconfirm-content-pane' ),
+					rootMargin: '600px', // 5 rows of icons. Formula: 20 + ( (96 + 20) * rows ).
+				};
+
+				let observer = new IntersectionObserver( function( entries ) {
+
+					if ( ! entries[0].isIntersecting ) {
+						return;
+					}
+
+					page++;
+					list.show( 0, page * wpforms_builder.icon_choices.icons_per_page );
+				}, options );
+
+				observer.observe( document.querySelector( '.wpforms-icon-picker-pagination' ) );
+			},
+
+			/**
+			 * When an icon is selected, update the choice and the field preview.
+			 *
+			 * @since 1.7.9
+			 *
+			 * @param {object} modal Current jQuery Confirm modal instance.
+			 * @param {jquery} $this The list item (icon) that was clicked.
+			 */
+			selectIcon: function( modal, $this ) {
+
+				const fieldId = $this.parent().data( 'field-id' );
+				const choiceId = $this.parent().data( 'choice-id' );
+				const icon = $this.data( 'icon' );
+				const iconStyle = $this.data( 'icon-style' );
+				const $choice = $( '#wpforms-field-option-row-' + fieldId + '-choices ul li[data-key=' + choiceId + ']' );
+				const fieldType = $( '#wpforms-field-option-row-' + fieldId + '-choices ul' ).data( 'field-type' );
+
+				$this.addClass( 'selected' );
+				$this.siblings( '.selected' ).removeClass( 'selected' );
+
+				$choice.find( '.wpforms-icon-select span' ).text( icon );
+				$choice.find( '.wpforms-icon-select .ic-fa-preview' ).removeClass().addClass( `ic-fa-preview ic-fa-${iconStyle} ic-fa-${icon}` );
+				$choice.find( '.wpforms-icon-select .source-icon' ).val( icon );
+				$choice.find( '.wpforms-icon-select .source-icon-style' ).val( iconStyle );
+
+				app.fieldChoiceUpdate( fieldType, fieldId );
+
+				modal.close();
+			},
+		},
+
+		//--------------------------------------------------------------------//
 		// Alerts (notices).
 		//--------------------------------------------------------------------//
+
 		/**
 		 * Click on the Dismiss notice button.
 		 *
@@ -7337,9 +8006,46 @@ var WPFormsBuilder = window.WPFormsBuilder || ( function( document, window, $ ) 
 		 * Load or refresh color picker.
 		 *
 		 * @since 1.2.1
+		 * @since 1.7.9 Added default value support.
 		 */
 		loadColorPickers: function() {
-			$( '.wpforms-color-picker' ).minicolors();
+
+			$( '.wpforms-color-picker' ).each( function() {
+
+				const $this = $( this );
+
+				// If it appears to be already initialized, reset. This is needed when duplicating fields with color pickers.
+				if ( $this.hasClass( 'minicolors-input' ) ) {
+					$this.minicolors( 'destroy' );
+				}
+
+				$this.minicolors( {
+					defaultValue: $this.data( 'fallback-color' ) || '',
+				} );
+			} );
+		},
+
+		/**
+		 * Get a valid color value from color picker or a default one.
+		 *
+		 * @since 1.7.9
+		 *
+		 * @param {object} $colorPicker Current field.
+		 *
+		 * @returns {string} Always valid color value.
+		 */
+		getValidColorPickerValue: function( $colorPicker ) {
+
+			const color = $colorPicker.minicolors( 'value' );
+
+			// jQuery MiniColors returns "black" RGB object if the color value is invalid.
+			const isInvalid = _.isEqual( $colorPicker.minicolors( 'rgbObject' ), { r: 0, g: 0, b: 0 } );
+			const isBlack = _.includes( [ '#000', '#000000' ], color );
+
+			// If default value isn't provided via the data attribute, use black.
+			const defaultValue = $colorPicker.data( 'fallback-color' ) || '#000000';
+
+			return isInvalid && ! isBlack ? defaultValue : color;
 		},
 
 		/**