+++ wp-content/plugins/wpforms/assets/js/components/admin/builder/setup.js
@@ -445,7 +445,10 @@ WPForms.Admin.Builder.Setup = WPForms.Admin.Builder.Setup || ( function( documen
 
 						// We have already warned the user that unsaved changes will be ignored.
 						WPFormsBuilder.setCloseConfirmation( false );
-						window.location.href = res.data.redirect;
+
+						window.location.href = wpf.getQueryString( 'force_desktop_view' ) ?
+							wpf.updateQueryString( 'force_desktop_view', 1, res.data.redirect ) :
+							res.data.redirect;
 
 						return;
 					}