+++ wp-content/plugins/wpforms/CHANGELOG.md
@@ -1,6 +1,42 @@
 # Changelog
 All notable changes to this project will be documented in this file and formatted via [this recommendation](https://keepachangelog.com/).
 
+## [1.7.9] - 2023-01-03
+### Added
+- Icon Choices feature for Checkboxes, Multiple Choice, Checkbox Items, and Multiple Items payment fields - a selection of 2000+ icons can now be used with your choices!
+
+### Changed
+- Avoid rendering the WPForms Import admin page if the user lacks `unfiltered_html` capability.
+- Respect site settings for displaying avatars on the Revisions screen in the Form Builder.
+- Minor CSS adjustments on the Entry details page.
+- Color picker fields in the Form Builder are now correctly handling default colors.
+- In the form Notifications you can now set up the Reply-To Name value in addition to the Reply-To Email using a special format.
+- Updated jquery-confirm library to 3.3.4.
+
+### Fixed
+- Improved a preview for the Classic File Upload field in the Form Builder.
+- Prevent field duplication in the Form Builder performed multiple times when clicking fast inside the confirmation modal.
+- Action links were rendered on two lines in the admin dashboard widget.
+- The content editor option in HTML mode was not visible when the Content field was added inside the Layout field.
+- Buttons of the content editor option in Visual mode didn't have hotkey texts in their tooltips when the Content field was added.
+- Email notification was not able to show the submitted Content field value.
+- Text and image styles were not applied to the Content field value on the Entry Print Preview page.
+- WordPress VIP platform users were unable to export form entries.
+- Styles for the Content field editor were not applied when the field had been placed in the Layout field in Firefox.
+- The content field disappeared if it was duplicated inside the Layout field.
+- A blank space was showing when all of the fields inside of a Layout field were hidden using the Conditional Logic.
+- PHP notices were generated when form locations logic ran for unregistered post types.
+- Not all WPForms-specific data was removed from the database when the Settings > Misc > Uninstall option was enabled.
+- The Embed modal performance in the Form Builder was improved a lot when there are a ton of pages on a site.
+- The confirmation message for non-AJAX form submissions wasn't wrapped into the main form container.
+- In the Layout field its last column on the right side was always wider than other columns.
+- Improved compatibility with Elementor popups v3.9+.
+- Notification email suggestion didn't work properly in WordPress installed in a subdomain.
+- License key was incorrectly processed when set in the `wp-config.php`.
+- `{page_title}` smart tag was conflicting with the wpSEO plugin.
+- Better compatibility with the Popup Maker plugin.
+- Activate the first form page with an error after failed form submission for AJAX forms.
+
 ## [1.7.8] - 2022-11-09
 ### Added
 - Introducing a completely new Content field to help you easily add formatted text to your forms.